+++ src/completer.js
@@ -57,6 +57,10 @@
     };
   };
 
+  var isString = function (obj) {
+    return Object.prototype.toString.call(obj) === '[object String]';
+  };
+
   var uniqueId = 0;
 
   function Completer(element, option) {
@@ -166,8 +170,12 @@
     _extractSearchQuery: function (text) {
       for (var i = 0; i < this.strategies.length; i++) {
         var strategy = this.strategies[i];
-        var match = text.match(strategy.match);
-        if (match) { return [strategy, match[strategy.index]]; }
+        var context = strategy.context(text);
+        if (context || context === '') {
+          if (isString(context)) { text = context; }
+          var match = text.match(strategy.match);
+          if (match) { return [strategy, match[strategy.index]]; }
+        }
       }
       return []
     },